Split the PDF bleach into smaller classes
Each class is responsible for one part of the PDF:
- COSObjectBleach handles raw PDF objects, containing
  potential JavaScript code.
- PDAnnotationBleach handles annotations, as they may
  contain JavaScript code.
- PDDocumentCatalogBleach handles the document catalog,
  which includes the pages, the forms and their respective
  actions.
- PDEmbeddedFileBleach handles embedded files.

PdfBleachSession is a helper class to store a shared state.
This design may not be perfect yet, but it has several advantages:
- BleachSession is no longer a parameter to every method
- Test files target one specific part of the document

The test for the annotations (sanitizeAnnotationWidgetActions) was
rewritten using reflection. The reflection may be pushed even further
to detect new methods added in PDFBox, and in the class to
automatically use them - at the cost of not catching these bugs at
compile time.